integration: test ready state
