+++ integration/api_tests/test_likes.py
@@ -1,8 +1,9 @@
-from config import url
-
+from framework.config import api
 import httpx
+import pytest
 
 
+@pytest.mark.skip(reason="not implemented")
 def tests_post_can_be_liked():
     headers = {"Content-Type": "application/json", "Authorization": "Bearer 123"}
     like = {
@@ -10,12 +11,13 @@ def tests_post_can_be_liked():
         "profileId": "e2f0fa7e-4517-4ac8-bbc6-73067d3feed4",
     }
 
-    r = httpx.post(url + "/likes", json=like, headers=headers)
+    r = httpx.post(api + "/likes", json=like, headers=headers)
 
     assert r.status_code == 200
     assert r.json() == like
 
 
+@pytest.mark.skip(reason="not implemented")
 def tests_post_can_be_unliked():
     headers = {"Content-Type": "application/json", "Authorization": "Bearer 123"}
     like = {
@@ -24,7 +26,7 @@ def tests_post_can_be_unliked():
     }
 
     r = httpx.delete(
-        url + f"/likes/{like['postId']}/{like['profileId']}", headers=headers
+        api + f"/likes/{like['postId']}/{like['profileId']}", headers=headers
     )
 
     assert r.status_code == 200
