+++ integration/api_tests/test_profiles.py
@@ -1,12 +1,13 @@
-from config import url
+from framework.config import api
 
 import httpx
 import pytest
 
 path = "/profiles"
-route = url + path
+route = api + path
 
 
+@pytest.mark.skip(reason="Not implemented")
 def test_get_profile_by_username():
     r = httpx.get(route + "/t-eckert")
 
@@ -42,6 +43,7 @@ def test_get_blogs_by_username():
     assert r.json() == expected
 
 
+@pytest.mark.skip(reason="Not implemented")
 def test_get_all_liked_post_ids_for_profile():
     r = httpx.get(route + "/e2f0fa7e-4517-4ac8-bbc6-73067d3feed4/likes/ids")
     print(r.json())
