+++ integration/api_tests/test_ready.py
@@ -1,11 +1,7 @@
-from config import url
-
+from framework.config import api
 import httpx
-
-path = "/ready"
-route = url + path
 
 
 def test_ready_route_returns_ready():
-    r = httpx.get(route)
+    r = httpx.get(api + "/ready")
     assert r.status_code == 200
