+++ integration/framework/utils.py
@@ -1,0 +1,4 @@
+def is_subset(a: dict, b: dict) -> bool:
+    # Iterate through keys in a and check key/value equality in b,
+    # ignoring keys not in a.
+    return all(a[k] == b[k] for k in a if k in b)
